+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+Unreleased
+----------
+
+**Added**
+
+* :attr:`.Redditor.stream`, with methods :meth:`.RedditorStream.submissions()`
+  and :meth:`.RedditorStream.comments()` to stream a Redditor's
+  comments or submissions
+* :class:`.Redditor` has been added to facilitate
+  :attr:`.Redditor.stream`
+
+
 5.0.1 (2017/07/11)
 ------------------